Refactoring Runtime.Api - SuggestionAction
Moving `ExportsAction` out of the `Runtime` class.
